requirements.txt
django >= 1.7
djangorestframework >= 2.4.3
gunicorn >= 19.1.1
django-post-office >= 1.1.1
django-cors-headers >= 0.13
django-celery-with-redis
django-grappelli
django-sslserver
django-model-utils
psycopg2